Harden generated live setting numeric literals
Reject non-finite double live setting defaults during source generation instead of emitting invalid generated literals. Also preserve valid long RangeAttribute values that arrive through the numeric double overload beyond Int32 bounds.\n\nThis improves generator completeness by covering supported manifest numeric shapes and keeps unsupported finite-value cases as explicit SGP100 diagnostics.

+++ src/SafeIR.PluginAnalyzer/PluginSymbolReader.cs
@@ -236,7 +236,7 @@ internal static class PluginSymbolReader
         {
             int number => number,
             long number => number,
-            double number when IsWhole(number) && number >= int.MinValue && number <= int.MaxValue => (long)number,
+            double number when IsWhole(number) && number >= long.MinValue && number <= long.MaxValue => (long)number,
             string text => long.Parse(text, NumberStyles.Integer, CultureInfo.InvariantCulture),
             _ => throw RangeValueException()
         };
